Fixed a bug that caused a load-data-trigger to fail to trip whenever another trigger (e.g. instruction address or opcode) has tripped.
Fixed a bug that allowed a performance counter to count even
when that counter was the target of a csr-write instruction.

Instruction fetch to PIC addresses will now go to external memory!
This was done to match the RTL.

Initial memory load (from ELF or HEX file) will now put data in a
memory mapped register and the external memory with the same address
to match test-bench.

Memory mapped regiters (PIC) region is no longer affected by memory
access windows.