Fixed syncrhonization bug in implementation of LR instruction.
Address translation work in progress: Translation hooked in for ld/st
and instsruction fetch. Not yet tested.

Consider PMP enabled if any of the PMP address registers is
implemented. Before we considered PMP enabled if any of the PMP
configs was not off.

Now supporting memory mapped software interrupt.

Fixed bug that cause masked triggers to mismatch when the least
significant zero bit of tdata2 was at position 31 (in a 32-bit
machine). C++ code shifting by 32 bits was being ingored by the
compiler.

Fixed by in PMP code: In NAPOT mode, if PMPADDR was all ones, the size
(2 to the power xlen+3) was overflowing.

In server mode, we now log commands that are incorrectly directed to a
non-started hart.

Pokes will now succeed in PIC region after region is relocated.

A precise error from the bench targeting external memory will now
cause such memory to be written.

Fixed bug that cause masked triggers to mismatch when the least
significant zero bit of tdata2 was at position 31 (in a 32-bit
machine).  C++ code shifting by 32 bits was being ingored by the
compiler.

Fixed a bug that caused a load-data-trigger to trip in the presence of
an exception.

Now supporting memory mapped software interrupt with the new
--softinterrupt option.