Fixed bug in csrci: Immediate was not being extended to 64-bits in rv64.
Fixed bug in the mask of sstatus that caused the uxl field to be
incorrect in rv64.

Fixed bug that made access to the satp CSR legal in supervisor mode
when cstatus.tvm was 1.

We now check PMP after an instruction address trigger trips.

Reading CSR SATP now triggers an illegal instruction exception if in
supervisor mode and MSTATUS.TVM is 1.